bot/requirements.txt
aiogram==3.26.0
aiosqlite==0.20.0
aiohttp==3.11.0
python-dotenv==1.0.1
Pillow>=10.0.0
